Adding basic validation of input parameters for signature forms, and refactored key pieces into controller helper.
